DDF-3875 Build admin-federation-ui with ace instead of grunt
