+++ ui/packages/admin-federation-ui/package.json
@@ -4,10 +4,6 @@
   "description": "A frontend UI for federation configuration on DDF.",
   "version": "0.1.0",
   "license": "LGPL-3.0",
-  "keywords": [
-    "backbone",
-    "express"
-  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/codice/ddf.git"
@@ -16,41 +12,29 @@
     "node": ">=0.10.5"
   },
   "devDependencies": {
-    "colors": "1.1.2",
-    "connect-livereload": "0.3.2",
-    "grunt": "1.0.1",
-    "grunt-cli": "1.2.0",
-    "grunt-contrib-clean": "1.0.0",
-    "grunt-contrib-cssmin": "1.0.2",
-    "grunt-contrib-jshint": "1.1.0",
-    "grunt-contrib-less": "1.4.0",
-    "grunt-contrib-watch": "1.0.0",
-    "grunt-express-server": "0.5.3",
-    "grunt-port-allocator": "*",
-    "jshint-stylish": "1.0.2",
-    "load-grunt-tasks": "3.5.2",
-    "which": "1.0.5",
-    "yargs": "1.3.3",
     "ace": "*"
   },
   "dependencies": {
-    "express": "4.16.3",
-    "faye": "1.1.0",
-    "lodash": "4.17.4",
-    "node-fs": "0.1.7",
-    "node-options": "0.0.3",
-    "path": "0.4.9",
-    "request": "2.85.0"
+    "backbone": "1.1.2",
+    "backbone-associations": "0.6.2",
+    "backbone.marionette": "2.4.7",
+    "jquery": "3.2.1"
+  },
+  "alias": {
+    "marionette": "backbone.marionette"
   },
   "scripts": {
-    "start": "grunt",
-    "build": "grunt build && ace package",
-    "clean": "ace clean"
+    "startplus": "ace start --open",
+    "start": "ace start",
+    "pretest": "ace bundle --env=test",
+    "test": "ace test ./target/test/index.html",
+    "testplus": "ace start --env=test --open",
+    "build": "ace bundle",
+    "postbuild": "ace package"
   },
   "files": [
-    "src/main/webapp",
-    "src/main/resources",
     "target/webapp"
   ],
+  "main": "src/main/webapp/js/main.js",
   "context-path": "/admin/federation"
 }
