+++ ui/packages/admin-federation-ui/src/main/webapp/js/application.js
@@ -16,17 +16,21 @@ define([
     'underscore',
     'backbone',
     'marionette',
-    'icanhaz'
-    ],function (_, Backbone, Marionette, ich) {
+    'ace/handlebars'
+    ],function (_, Backbone, Marionette, hbs) {
     'use strict';
 
     var Application = {};
 
     // This was moved from the main.js file into here.
     // Since this modules has ui components, and it gets loaded before main.js, we need to init the renderer here for now until we sort this out.
-    Marionette.Renderer.render = function (template, data) {
-        if(!template){return '';}
-        return ich[template](data);
+    Marionette.Renderer.render = function(template, data, view) {
+        data._view = view;
+        if (typeof template === 'function') {
+            return template(data);
+        } else {
+            return hbs.compile(template)(data);
+        }
     };
 
     Application.App = new Marionette.Application();
