+++ ui/packages/admin-federation-ui/src/main/webapp/js/model/Federation.js
@@ -13,7 +13,7 @@
 
 define([
         'backbone',
-        'backboneassociation'
+        'backbone-associations'
   ],
   function (Backbone) {
     "use strict";
