+++ ui/packages/admin-federation-ui/src/main/webapp/js/module.js
@@ -51,6 +51,5 @@ define([
                 FederationModule.contentController.show();
             });
 
-
         });
     });
