+++ ui/packages/admin-federation-ui/src/main/webapp/styles/styles.less
