+++ ui/packages/admin-federation-ui/src/main/webapp/templates/federationPage.handlebars
@@ -18,7 +18,7 @@
  {{/each}}
 </div>
 <div class="tab-content">
-  <iframe src="{{url}}" width="100%" scrolling="no" frameBorder="0">
+  <iframe src="{{url}}" width="100%" scrolling="no" frameBorder="0" onload="iFrameResize();">
     <p>Your browser does not support iframes.</p>
   </iframe>
 </div>
